% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_golem_config}
\alias{get_golem_config}
\title{Read App Config}
\usage{
get_golem_config(
  value,
  config = Sys.getenv("GOLEM_CONFIG_ACTIVE", Sys.getenv("R_CONFIG_ACTIVE", "default")),
  use_parent = TRUE,
  file = app_sys("inst/golem-config.yml")
)
}
\description{
Read App Config
}
